HYDRA-1057 : Move assertSelectionHighlightCorrectness to testUtils
